Create deprecated fncs to replace to-be-removed macros
These macros should not be used as they are prone to misuse.  There are
no occurrences of them in CPAN.  The single use of either of them in
core has recently been removed (commit
8d40577bdbdfa85ed3293f84bf26a313b1b92f55), because it was a misuse.

Instead code should use isIDFIRST_lazy_if or isWORDCHAR_lazy_if
(isALNUM_lazy_if is also available, but can be confused with the Posix
alnum, which it doesn't mean).